requirements.txt
# GUI framework
PySide6>=6.6.0

# LangChain ecosystem (RAG pipeline)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-ollama>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store
faiss-cpu>=1.7.0

# Document loaders
pypdf>=3.0.0
python-docx>=1.0.0

# Local embeddings (no API key required)
sentence-transformers>=3.0.0

# LLM provider SDKs
openai>=1.0.0
anthropic>=0.20.0

# Model Context Protocol (>=1.23.0 required to fix DNS-rebinding vulnerability)
mcp>=1.23.0
